Moved history component, added start draft of frontend code style guide
