+++ public/app/features/dashboard/components/VersionHistory/index.ts
@@ -1,0 +1,2 @@
+export { HistoryListCtrl } from './HistoryListCtrl';
+export { HistorySrv } from './HistorySrv';
