+++ public/app/features/dashboard/index.ts
@@ -1,6 +1,5 @@
 import './dashboard_ctrl';
 import './alerting_srv';
-import './history/history';
 import './dashboard_loader_srv';
 import './submenu/submenu';
 import './save_as_modal';
@@ -27,6 +26,7 @@ import './components/DashExportModal';
 import './components/DashNav';
 import './components/ExportDataModal';
 import './components/FolderPicker';
+import './components/VersionHistory';
 
 // angular wrappers
 import { react2AngularDirective } from 'app/core/utils/react2angular';
